Add files via upload
This is a curated 12S reference database that was created for Czachur et al. (in prep), with the intention of improving taxonomic assignments of fishes in South Africa and Mozambique. The fasta file, Czachur_DB_v3.fa, can be used directly with the MiFish metabarcoding pipeline (see Miya et al. 2020)